Feito implements e services faltantes para Matricula

+++ src/br/org/catolicasc/academia/service/MatriculaService.java
@@ -6,9 +6,9 @@ import br.org.catolicasc.academia.model.Aluno;
 import br.org.catolicasc.academia.model.Matricula;
 
 public interface MatriculaService {
-	Matricula findById(int id);
+	Matricula findById(Long id);
 	void save(Matricula matricula);
 	void update(Matricula matricula);
-	void delete(int id);
+	void delete(Long id);
 	List<Matricula> findAll();
 }
